# By Darren McDaniel (17) and Michael Sallaway (2)
# Via Darren McDaniel (1) and Michael Sallaway (1)
* gizmo-master-CRLF_fixups:
  Fix up line endings on Giz/master to CRLF on changed files since 2696124a5ccf32abd827096489a2655e63cfa0d8.
  Remove .gitattributes file causing line endings issues.
  Fix to AutomaticMusicDetection to prevent an exception from firing after a user has uses the tool and closes the mark manager form
  Updated Version Control filewatchers
  Remove references to VB2005 Powerpacks assembly
  Fix VersionCOntrol where it wasnt getting committed until after a 10second wait.. And, updated commits were not available to view until after the application restarted
  UPdated pulse
  Revert "First Stab at migrating many of the Double objects to Float's"
  Failed revert... made the app unbuildable.... Revert "Revert "First Stab at migrating many of the Double objects to Float's""
  Revert "Git insists these need committing even tho nothing has changed"
  Updated Version Control Module with Load Status indicator
  Revert "First Stab at migrating many of the Double objects to Float's"
  Git insists these need committing even tho nothing has changed
  Changed width on Audio Device dropdown
  Change the Editor toolbars from three separate to one combined to utilize the space
  Minor fixes to how NShape  build was configured.  Had to force NShape to ANyCPU and .NET 2.0  so that NShape modules would properly show in the designer
  Updated references
  Fixed reference and missing file
  Updated Installer

Conflicts:
	Modules/App/WebServer/HTTP/RequestDelegate.cs
	Modules/Editor/TimedSequenceEditor/TimedSequenceEditorForm.Designer.cs
	Modules/Editor/TimedSequenceEditor/TimedSequenceEditorForm.cs

+++ Common/Controls/ColorPicker/ColorSelectionModules_HSV.cs
@@ -13,14 +13,14 @@ namespace Common.Controls.ColorManagement.ColorPicker
 
 		#endregion
 
-		protected static ColorBlend GetHueBlend(float s, float v)
+		protected static ColorBlend GetHueBlend(double s, double v)
 		{
 			ColorBlend ret = new ColorBlend();
 			ret.Colors = new Color[]
 			             	{
-			             		new HSV(0.0f, s, v).ToRGB(), new HSV(0.1666f, s, v).ToRGB(), new HSV(0.3333f, s, v).ToRGB(),
-			             		new HSV(0.5f, s, v).ToRGB(), new HSV(0.6666f, s, v).ToRGB(), new HSV(0.8333f, s, v).ToRGB(),
-			             		new HSV(1.0f, s, v).ToRGB()
+			             		new HSV(0.0, s, v).ToRGB(), new HSV(0.1666, s, v).ToRGB(), new HSV(0.3333, s, v).ToRGB(),
+			             		new HSV(0.5, s, v).ToRGB(), new HSV(0.6666, s, v).ToRGB(), new HSV(0.8333, s, v).ToRGB(),
+			             		new HSV(1.0, s, v).ToRGB()
 			             	};
 			ret.Positions = new float[]
 			                	{
@@ -60,7 +60,7 @@ namespace Common.Controls.ColorManagement.ColorPicker
 				using (LinearGradientBrush brs = new LinearGradientBrush(
 					new Point(0, 0), new Point(bmp.Width),
 					Color.Black, Color.White)) {
-					brs.InterpolationColors = GetHueBlend(1.0f, 1.0f);
+					brs.InterpolationColors = GetHueBlend(1.0, 1.0);
 					gr.FillRectangle(brs, new Rectangle(Point.Empty, bmp.Size));
 				}
 			}
@@ -73,7 +73,7 @@ namespace Common.Controls.ColorManagement.ColorPicker
 
 		protected override void OnFaderScroll(ColorSelectionFader fader)
 		{
-			float newhue = fader.Position;
+			double newhue = fader.Position;
 			if (newhue == _color.H) return;
 			_color.H = newhue;
 			UpdatePlaneImage();
@@ -103,13 +103,13 @@ namespace Common.Controls.ColorManagement.ColorPicker
 
 		protected override void OnUpdatePlanePosition(ColorSelectionPlane plane)
 		{
-			plane.SetPosition(_color.S, 1.0f - _color.V);
+			plane.SetPosition(_color.S, 1.0 - _color.V);
 		}
 
 		protected override void OnPlaneScroll(ColorSelectionPlane plane)
 		{
-			float newsaturation = XYZ.ClipValue(plane.PositionX, 0.0f, 1.0f),
-			       newvalue = 1.0f - XYZ.ClipValue(plane.PositionY, 0.0f, 1.0f);
+			double newsaturation = XYZ.ClipValue(plane.PositionX, 0.0, 1.0),
+			       newvalue = 1.0 - XYZ.ClipValue(plane.PositionY, 0.0, 1.0);
 			if (newsaturation == _color.S &&
 			    newvalue == _color.V) return;
 			_color.S = newsaturation;
@@ -129,8 +129,8 @@ namespace Common.Controls.ColorManagement.ColorPicker
 			using (Graphics gr = Graphics.FromImage(bmp)) {
 				using (LinearGradientBrush brs = new LinearGradientBrush(
 					new Point(0, 0), new Point(bmp.Width),
-					new HSV(_color.H, 0.0f, _color.V).ToRGB(),
-					new HSV(_color.H, 1.0f, _color.V).ToRGB())) {
+					new HSV(_color.H, 0.0, _color.V).ToRGB(),
+					new HSV(_color.H, 1.0, _color.V).ToRGB())) {
 					gr.FillRectangle(brs, new Rectangle(Point.Empty, bmp.Size));
 				}
 			}
@@ -143,7 +143,7 @@ namespace Common.Controls.ColorManagement.ColorPicker
 
 		protected override void OnFaderScroll(ColorSelectionFader fader)
 		{
-			float newsaturation = fader.Position;
+			double newsaturation = fader.Position;
 			if (newsaturation == _color.S) return;
 			_color.S = newsaturation;
 			UpdatePlaneImage();
@@ -162,7 +162,7 @@ namespace Common.Controls.ColorManagement.ColorPicker
 					Color.White, Color.White)) {
 					//draw hue
 					brs.Transform = new Matrix((float) bmp.Width, 0f, 0f, 1f, 0f, 0f);
-					brs.InterpolationColors = GetHueBlend(_color.S, 1.0f);
+					brs.InterpolationColors = GetHueBlend(_color.S, 1.0);
 					gr.FillRectangle(brs, new Rectangle(Point.Empty, bmp.Size));
 					//draw value
 					brs.Transform = new Matrix(0f, (float) bmp.Height, 1f, 0f, 0f, 0f);
@@ -177,13 +177,13 @@ namespace Common.Controls.ColorManagement.ColorPicker
 
 		protected override void OnUpdatePlanePosition(ColorSelectionPlane plane)
 		{
-			plane.SetPosition(_color.H, 1.0f - _color.V);
+			plane.SetPosition(_color.H, 1.0 - _color.V);
 		}
 
 		protected override void OnPlaneScroll(ColorSelectionPlane plane)
 		{
-			float newhue = XYZ.ClipValue(plane.PositionX, 0.0f, 1.0f),
-			       newvalue = 1.0f - XYZ.ClipValue(plane.PositionY, 0.0f, 1.0f);
+			double newhue = XYZ.ClipValue(plane.PositionX, 0.0, 1.0),
+			       newvalue = 1.0 - XYZ.ClipValue(plane.PositionY, 0.0, 1.0);
 			if (newhue == _color.H &&
 			    newvalue == _color.V) return;
 			_color.H = newhue;
@@ -204,7 +204,7 @@ namespace Common.Controls.ColorManagement.ColorPicker
 			using (Graphics gr = Graphics.FromImage(bmp)) {
 				using (LinearGradientBrush brs = new LinearGradientBrush(
 					new Point(0, 0), new Point(bmp.Width),
-					Color.Black, new HSV(_color.H, _color.S, 1.0f).ToRGB()
+					Color.Black, new HSV(_color.H, _color.S, 1.0).ToRGB()
 					)) {
 					gr.FillRectangle(brs, new Rectangle(Point.Empty, bmp.Size));
 				}
@@ -218,7 +218,7 @@ namespace Common.Controls.ColorManagement.ColorPicker
 
 		protected override void OnFaderScroll(ColorSelectionFader fader)
 		{
-			float newvalue = fader.Position;
+			double newvalue = fader.Position;
 			if (newvalue == _color.V) return;
 			_color.V = newvalue;
 			UpdatePlaneImage();
@@ -237,12 +237,12 @@ namespace Common.Controls.ColorManagement.ColorPicker
 					Color.White, Color.White)) {
 					//draw hue
 					brs.Transform = new Matrix((float) bmp.Width, 0f, 0f, 1f, 0f, 0f);
-					brs.InterpolationColors = GetHueBlend(1.0f, _color.V);
+					brs.InterpolationColors = GetHueBlend(1.0, _color.V);
 					gr.FillRectangle(brs, new Rectangle(Point.Empty, bmp.Size));
 					//draw value
 					brs.Transform = new Matrix(0f, (float) bmp.Height, 1f, 0f, 0f, 0f);
 					ColorBlend blnd = new ColorBlend();
-					Color zerosat = new HSV(0.0f, 0.0f, _color.V).ToRGB();
+					Color zerosat = new HSV(0.0, 0.0, _color.V).ToRGB();
 					blnd.Colors = new Color[] {Color.FromArgb(0, zerosat), zerosat};
 					blnd.Positions = new float[] {0f, 1f};
 					brs.InterpolationColors = blnd;
@@ -253,13 +253,13 @@ namespace Common.Controls.ColorManagement.ColorPicker
 
 		protected override void OnUpdatePlanePosition(ColorSelectionPlane plane)
 		{
-			plane.SetPosition(_color.H, 1.0f - _color.S);
+			plane.SetPosition(_color.H, 1.0 - _color.S);
 		}
 
 		protected override void OnPlaneScroll(ColorSelectionPlane plane)
 		{
-			float newhue = XYZ.ClipValue(plane.PositionX, 0.0f, 1.0f),
-			       newsaturation = 1.0f - XYZ.ClipValue(plane.PositionY, 0.0f, 1.0f);
+			double newhue = XYZ.ClipValue(plane.PositionX, 0.0, 1.0),
+			       newsaturation = 1.0 - XYZ.ClipValue(plane.PositionY, 0.0, 1.0);
 			if (newhue == _color.H &&
 			    newsaturation == _color.S) return;
 			_color.H = newhue;
