+++ Common/Controls/ColorPicker/ColorSelectionModules_RGB.cs
@@ -52,7 +52,7 @@ namespace Common.Controls.ColorManagement.ColorPicker
 
 		protected override void OnUpdateFaderPosition(ColorSelectionFader fader)
 		{
-			fader.Position = _color.R / 255.0f;
+			fader.Position = (double) (_color.R)/255.0;
 		}
 
 		protected override void OnFaderScroll(ColorSelectionFader fader)
@@ -91,7 +91,8 @@ namespace Common.Controls.ColorManagement.ColorPicker
 
 		protected override void OnUpdatePlanePosition(ColorSelectionPlane plane)
 		{
-			plane.SetPosition(((float)_color.B) / 255.0f, 1.0f - ((float)_color.G) / 255.0f);
+			plane.SetPosition((double) (_color.B)/255.0,
+			                  1.0 - (double) (_color.G)/255.0);
 		}
 
 		protected override void OnPlaneScroll(ColorSelectionPlane plane)
@@ -126,7 +127,7 @@ namespace Common.Controls.ColorManagement.ColorPicker
 
 		protected override void OnUpdateFaderPosition(ColorSelectionFader fader)
 		{
-			fader.Position = _color.G / 255.0f;
+			fader.Position = (double) (_color.G)/255.0;
 		}
 
 		protected override void OnFaderScroll(ColorSelectionFader fader)
@@ -165,7 +166,8 @@ namespace Common.Controls.ColorManagement.ColorPicker
 
 		protected override void OnUpdatePlanePosition(ColorSelectionPlane plane)
 		{
-			plane.SetPosition(((float)_color.B) / 255.0f, 1.0f - ((float)_color.R) / 255.0f);
+			plane.SetPosition((double) (_color.B)/255.0,
+			                  1.0 - (double) (_color.R)/255.0);
 		}
 
 		protected override void OnPlaneScroll(ColorSelectionPlane plane)
@@ -200,7 +202,7 @@ namespace Common.Controls.ColorManagement.ColorPicker
 
 		protected override void OnUpdateFaderPosition(ColorSelectionFader fader)
 		{
-			fader.Position = _color.B / 255.0f;
+			fader.Position = (double) (_color.B)/255.0;
 		}
 
 		protected override void OnFaderScroll(ColorSelectionFader fader)
@@ -239,7 +241,8 @@ namespace Common.Controls.ColorManagement.ColorPicker
 
 		protected override void OnUpdatePlanePosition(ColorSelectionPlane plane)
 		{
-			plane.SetPosition(_color.R / 255.0f, 1.0f - _color.G / 255.0f);
+			plane.SetPosition((double) (_color.R)/255.0,
+			                  1.0 - (double) (_color.G)/255.0);
 		}
 
 		protected override void OnPlaneScroll(ColorSelectionPlane plane)
