+++ Modules/Effect/Pulse/Pulse.cs
@@ -180,8 +180,8 @@ namespace VixenModules.Effect.Pulse
 			foreach (PointPair point in LevelCurve.Points) {
 				points.Add(point.X/100);
 			}
-			float lastPointPos = 0.0f;
-			float lastDistinctPos = 0.0f;
+			double lastPointPos = 0.0;
+			double lastDistinctPos = 0.0;
 			bool addNextPointAsFadeOut = false;
 			foreach (ColorPoint point in ColorGradient.Colors.SortedArray()) {
 				if (color != null) {
