+++ Modules/Effect/Twinkle/Twinkle.cs
@@ -322,8 +322,8 @@ namespace VixenModules.Effect.Twinkle
 							break;
 
 						case TwinkleColorHandling.GradientThroughWholeEffect:
-							float startPos = ((float)twinkle.StartTime.Ticks / (float)TimeSpan.Ticks);
-							float endPos = ((float)(twinkle.StartTime + twinkle.Duration).Ticks / (float)TimeSpan.Ticks);
+							double startPos = ((double) twinkle.StartTime.Ticks/(double) TimeSpan.Ticks);
+							double endPos = ((double) (twinkle.StartTime + twinkle.Duration).Ticks/(double) TimeSpan.Ticks);
 
 							if (discreteColors) {
 								double range = endPos - startPos;
