+++ Vixen.System/Interpolator/Interpolator.cs
@@ -11,11 +11,11 @@ namespace Vixen.Interpolator
 	{
 		public bool Interpolate(TimeSpan timeOffset, TimeSpan timeSpan, T startValue, T endValue, out T value)
 		{
-			float percent = (float) timeOffset.Ticks/timeSpan.Ticks;
+			double percent = (double) timeOffset.Ticks/timeSpan.Ticks;
 			return Interpolate(percent, startValue, endValue, out value);
 		}
 
-		public bool Interpolate(float percentage, T startValue, T endValue, out T value)
+		public bool Interpolate(double percentage, T startValue, T endValue, out T value)
 		{
 			value = default(T);
 
@@ -30,7 +30,7 @@ namespace Vixen.Interpolator
 			return false;
 		}
 
-		protected abstract T InterpolateValue(float percent, T startValue, T endValue);
+		protected abstract T InterpolateValue(double percent, T startValue, T endValue);
 	}
 
 	public static class Interpolator
